+++ modules/kogito-data-index-infinispan/added/kogito-app-launch.sh
@@ -26,4 +26,4 @@ exec java ${SHOW_JVM_SETTINGS} ${DYNAMIC_RESOURCES_OPTS} ${JAVA_OPTIONS} ${KOGIT
         -Djava.library.path="${KOGITO_HOME}"/lib \
         -Dquarkus.http.host=0.0.0.0 \
         -Dquarkus.http.port=8080 \
-        -jar "${KOGITO_HOME}"/bin/data-index-service-infinispan-runner.jar
+        -jar "${KOGITO_HOME}"/bin/quarkus-app/quarkus-run.jar