+++ modules/kogito-explainability/added/kogito-app-launch.sh
@@ -11,7 +11,6 @@ if [ "${SCRIPT_DEBUG}" = "true" ] ; then
     printenv
 fi
 
-
 # Configuration scripts
 # Any configuration script that needs to run on image startup must be added here.
 CONFIGURE_SCRIPTS=(
@@ -24,8 +23,8 @@ source "${KOGITO_HOME}"/launch/configure.sh
 DYNAMIC_RESOURCES_OPTS="$(${JBOSS_CONTAINER_JAVA_JVM_MODULE}/java-default-options) $(${JBOSS_CONTAINER_JAVA_JVM_MODULE}/debug-options)"
 
 # shellcheck disable=SC2086
-exec java ${SHOW_JVM_SETTINGS} ${DYNAMIC_RESOURCES_OPTS} ${JAVA_OPTIONS} ${KOGITO_EXPLAINABILITY_PROPS} ${CUSTOM_TRUSTSTORE_ARGS} \
+exec java ${SHOW_JVM_SETTINGS} ${DYNAMIC_RESOURCES_OPTS} ${JAVA_OPTIONS} ${CUSTOM_TRUSTSTORE_ARGS} \
         -Djava.library.path="${KOGITO_HOME}"/lib \
         -Dquarkus.http.host=0.0.0.0 \
         -Dquarkus.http.port=8080 \
-        -jar "${KOGITO_HOME}"/bin/"${EXPLAINABILITY_SERVICE_JAR}"
+        -jar "${KOGITO_HOME}"/bin/"${EXPLAINABILITY_SERVICE_COMMUNICATION}"/quarkus-app/quarkus-run.jar