+++ modules/kogito-jit-runner/added/kogito-app-launch.sh
@@ -22,8 +22,8 @@ source "${KOGITO_HOME}"/launch/configure.sh
 DYNAMIC_RESOURCES_OPTS="$(${JBOSS_CONTAINER_JAVA_JVM_MODULE}/java-default-options) $(${JBOSS_CONTAINER_JAVA_JVM_MODULE}/debug-options)"
 
 # shellcheck disable=SC2086
-exec java ${SHOW_JVM_SETTINGS} ${DYNAMIC_RESOURCES_OPTS} ${JAVA_OPTIONS} ${KOGITO_JIT_RUNNER_PROPS} ${CUSTOM_TRUSTSTORE_ARGS} \
+exec java ${SHOW_JVM_SETTINGS} ${DYNAMIC_RESOURCES_OPTS} ${JAVA_OPTIONS} ${CUSTOM_TRUSTSTORE_ARGS} \
         -Djava.library.path="${KOGITO_HOME}"/lib \
         -Dquarkus.http.host=0.0.0.0 \
         -Dquarkus.http.port=8080 \
-        -jar "${KOGITO_HOME}"/bin/jitexecutor-runner.jar
+        -jar "${KOGITO_HOME}"/bin/quarkus-app/quarkus-run.jar